refactor into components, add some state
