+++ rqse/ch13/src/components/Button.jsx
@@ -1,0 +1,9 @@
+function Button({icon, title}) {
+  return(
+    <button title={title} className="icon-button">
+      <img src={`/icons/${icon}.svg`} alt={title} />
+    </button>
+  );
+}
+
+export default Button;
