+++ rqse/ch13/src/components/Header.jsx
@@ -1,0 +1,9 @@
+function Header({title}) {
+  return(
+    <header className="card-header">
+      <p className="card-title">{title}</p>
+    </header>
+  );
+}
+
+export default Header;
